+++ ADOL-C/include/adolc/interfaces.h
@@ -449,7 +449,7 @@ ADOLC_API int int_reverse_safe(short, int, int, int, const size_t *const *,
                                size_t **);
 
 /*--------------------------------------------------------------------------*/
-ADOLC_API int get_num_switches(short);
+ADOLC_API size_t get_num_switches(short);
 ADOLC_API int zos_pl_forward(short, int, int, int, const double *, double *,
                              double *);
 ADOLC_API short firstsign(int, const double *, const double *);
